requirements.txt
fastapi
uvicorn
sqlalchemy
sqlalchemy-utils
psycopg2
python-dotenv
fastapi[standard]